pyproject.toml
[project]
name = "long-memory-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.2",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "pgvector>=0.4.0",
    "psycopg2>=2.9.10",
    "psycopg[binary]>=3.2.6",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.1",
    "psycopg-pool>=3.0.0",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.14",
    "openai>=1.75.0",
]

[tool.uv.workspace]
members = [
    "client",
]
